refactor: extract EP combine-ready handler
